refactor(tests): 本地审查修复，共享设施口径与闸门一致
- CONTRIBUTING DB fixture 一条按代码现状重写：file_session_factory 带 uses_db
  但恒为文件 SQLite，补上 file_db_factory，登记 async_session PG 分支这一唯一
  不经 test_engine 的例外
- 过渡期说明收窄为「audit_tests.py 的 --check 形态尚不存在」
- 删 sent_json（无消费方）、make_test_audio 移回唯一消费方文件，两者都撞共享
  设施「公开符号须被 ≥2 个测试文件使用」闸门
- audit_tests.py 的 CONFTEST-IMPORT 覆盖 conftest 自身，补上「conftest 之间不
  互相 import」这条已声明却检不出的闸门
- dependabot dev-dependencies 分组补 respx
- 清理触及文件里的 issue 编号与时间性措辞注释
