+++ tests/factories.py
@@ -90,26 +90,6 @@ def make_test_video_with_audio_tail(
     )
 
 
-def make_test_audio(path: Path, *, duration_sec: float = 1.0) -> None:
-    """使用 ffmpeg 生成极短测试音频（正弦波 wav，pcm_s16le 为 ffmpeg 内置编码器）"""
-    path.parent.mkdir(parents=True, exist_ok=True)
-    subprocess.run(
-        [
-            "ffmpeg",
-            "-y",
-            "-f",
-            "lavfi",
-            "-i",
-            f"sine=frequency=440:duration={duration_sec}",
-            "-c:a",
-            "pcm_s16le",
-            str(path),
-        ],
-        capture_output=True,
-        check=True,
-    )
-
-
 def make_session_meta(**overrides) -> SessionMeta:
     """Build a SessionMeta with sensible defaults.
